stream::select_all: replace "use_std" feature check with "std" check
This also moves the select_all fn into select_all.rs,
since that's where it should be (this uses if_std to
check for std, and prevents further issues with
not updating a feature name)
